feat: add support for void type and dynamic memory allocation (#106)
* mugenjo

* few changes

* semantic init

* init testing

* cleanup?

* add void

* mugenjo 2.0

* docs: chane

* comments

* docs: add corent cases

* docs: some more edge cases

* docs: repositioning

+++ scasm/scasm.hh
@@ -381,8 +381,8 @@ struct backendSymbol {
   bool isTopLevel;
   /* Use for Functions */
   bool isDefined;
-  /* use for array */
-  int size = 0;
+  /*use for array*/
+  long size = 0;
   int alignment = 0;
 };
